requirements.txt
PyQt5>=5.15.0
watchdog>=2.1.0
pyinsane2>=2.0.6
Pillow>=9.0.0

